// interfaces/character.ts
export interface Character {
    id: number;
    name: string;
    description: string;
    modified: number;
    resourceURI: string;
    urls: [
        {
            type: string;
            url: string;
        }
    ];
    thumbnail: {
        path: string;
        extension: string;
    };
    comics: {
        available: number;
        returned: number;
        collectionURI: string;
        items: [
            {
                resourceURI: string;
                name: string;
            }
        ];
    };
    stories: {
        available: number;
        returned: number;
        collectionURI: string;
        items: [
            {
                resourceURI: string;
                name: string;
                type: string;
            }
        ];
    };
    events: {
        available: number;
        returned: number;
        collectionURI: string;
        items: [
            {
                resourceURI: string;
                name: string;
            }
        ];
    };
    series: {
        available: number;
        returned: number;
        collectionURI: string;
        items: [
            {
                resourceURI: string;
                name: string;
            }
        ];
    };
}